Clean up PHASER, REVPHASER, and MUSIC RIP effects
These were terrible when played using xptone and are now less
terrible.

This adds a new sweep() function that generates a sweep tone, and
adds a new WAVE_SHAPE_NO_CLEAN option to shapes which prevents
cleaning up the end of a waveform and extending a wave to a full
cycle.