Better error reporting in logonstats()
Eastside75 (THUNDER) is reporting that their daily statistics (ctrl/dsts.ini)
file is never "rolled-over" and there are only a couple of potential errors
in this function that could "silently fail" if the current date/time or the
"Date" value in the dsts.ini file could not be converted to a broken-down
date/time. So if either of those errors happen, log an error. It's unknown
what the cause of Eastside75's issue is, but I wanted to at least eliminate
these 2 possible errors (failures in calls to localtime_r()) as potential
causes.

Also, cleaned up a little using new dates_are_same() function, etc.