Support exceptions for saving log info of web, ftp, and services users
To reduce thrash on the userbase, it's now possible to specify which users
of some servers (web, ftp, and services) will have their user record
updated for every login. Normally, every user's login info (IP address,
hostname, protocol, date/time) is updated for every login to every
server. But for user accounts that are expected to have frequent logins
(e.g. Guest), this can be a source of excessive opening/locking/reading and
writing acitvity on the user base (data/user.tab file). So, if needed, a
sysop can limit which users login-info is updated for every login by setting
LoginInfoSave to an ARS value (e.g. "NOT GUEST") in the [Web], [Ftp], or
[Services] section of their ctrl/sbbs.ini file.

This is a mitigation specifically for Vertrauen where multiple instances of
SBBS share data files on a single Samba server and high "guest" activity leads
to occasional bursts of errors locking, reading, or writing the user.tab file.
This is likely not a feature that any other sysop will need or want to use.