Hang and scroll up fixes
This fixes ticket 207 by fixing the partial scrollup optimization
in bitmap output that scrolls down the screen, the scrolls up the
part that should not have moved.

While testing this change however, I managed to reproduce two unlikely
but possible deadlocks since I was rebuilding all the software on
my computer at the time and it was under very heavy load.

One of these issues was apparently found by Coverity and explcitly
suppressed... possibly this suppressed the other as a side-effect
as well.