Even with graphics, we need to set mouse passthrough
This used to work for a lot of complicated reasons you likely don't
care about, and the ANSI terminal overhaul broke it.

Basically, now, the only way to get mouse sequences passed to scripts
is to set the CON_MOUSE_*_PASSTHROUGH bits in console.status. The other
ways to do that don't work anymore, so don't try them.