More BBS list loading optimizations
When loading the BBS list, keep the list sorted and use bsearch()
to check for duplicates. Previously, they were unsorted and a case
insensitive string compare was performed between the new item and
every other item in the list.

After the last round of optimizations here, there was as much time
being spent checking for duplicates as was being spent loading and
parsing the data.

Now the bulk of the time is once more spent in the iniGet* functions,
and it's not my fault if it's slow.