Add support for root sections to iniParse and friends
Needed by SyncTERM.

Note, the error handling of iniParseSections() has changed somewhat...
Previously, on an allocation failure, it would return whatever it
had managed to parse so far with no indication of failure.

Now, it returns NULL.

Previously, NULL would be returned if there were no sections.

Now, NULL will not be returned if there's any global comments, and
will not be returned if there's any key/value pairs in the global
section.