Implement OSC 10 ; ? ST and OSC 11 ; ? ST
Yet another program (gcc) now defaults to dark blue on black for some things.

This pisses me off.

Apparently in the world we live in, every different program needs to use ANSI,
and it all needs a special snowflake environment variable (usually
PROGNAME_COLORS) to tell it to not suck on a terminal where dark blue is dark
blue and the default background colour is black.

To be fair, the default background colour is unknown to the program, so it
can't not suck on some terminals without overriding the background itself,
which would be quite rude, and the default default background colour in XTerm
is white.

XTerm supports OSC 10 ; ? ST and OSC 11 ; ? ST to query the RGB values of the
foreground and background colours respectively.  This would allow programs to
actually know and as a result, and take away their excuse for sucking.

Hopefully at least someone does this and chooses default colours based in it.