Put in an opt-in and opt-out compile flag
Since CFLAGS is used for multithread, we can't put -DWITHOUT_CRYPTLIB
in CFLAGS, but we can put -DWITH_CRYPTLIB there.

So now, to get the crypto stuff, you need to define WITH_CRYPTLIB
AND not define WITHOUT_CRYPTLIB.

This should "fix" all the builds everywhere, and just leave Windows
not able to use the new crypty stuff.