Fixes for a couple Cryptlib issues
Stream ciphers can be detected as having a blocksize of 1.

The return value of CRYPT_CTXINFO_IVSIZE is simply the block size,
which in the case of ChaCha20 is wildly incorrect, so we simply
can't use that attribute.  Instead, store the IV size at the start
of the IV itself.

WARNING: THIS CHANGE MAKES THE NEW CODE UNABLE TO DECRYPT A FILE
ENCRYPTED WITH THE OLD CODE. DECRYPT YOUR LIST BEFORE UPGRADING!