Try to fix intermitted test failures
The theory here is that "something" is modifying the files while
ld.so is linking them resulting in the crash (this seems to match
with the FreeBSD cores I've looked at).

To try to prevent this, have tar never overwrite files that are
being extracted.

While I'm here, remove the sync calls that didn't fix the issue.