A new theory for intermittent bus error failures in CI
We put the full path into the binary as the first RPATH entry, which
means if a library still exists in the directory it was originally
built in (ie: hasn't been cleaned up for the next job yet), it will
use that version instead of the one we extract.

If it happens to be deleted while mapping, the bus error occurs.

Add a new option (NO_ABSOLUTE_RPATH) to only use ORIGIN and use it
for the imacted targets.