*NIX doesn't use a textmode
OD_TEXTMODE is for when a text mode of a graphics card is being
used and so is really only applicable to DOS.

Add a new OD_HEADLESS for *nix, which more accurately models what's
going on here.

This fixes a display issue in GAC WaHoo because the "personality"
in use has a two-line status bar, and a write to the last column
of row 23 caused the internal buffer to wrap.

This means it's likely still broken on DOS, but can be easily fixed
with a custom personality that only uses one row for the status line.

I won't do that though.