Add support for a single-arg commandline callback
Previously, if you used the callback to parse unhandled arguments,
it would always eat two parameters from the command-line, which
means you simply couldn't use it for any existing doors.

The new od_cmd_line_flag_handler function will be passed a single
option, and if it returns TRUE, it will be assumed that it was
handled. If it returns FALSE, two options will be consumed, and
passed to the od_cmd_line_handler function.