Another overhaul... heap and overflows.
There was a lot of work keeping the stack small in favour of the
heap due to DOS limitations. Remove the obvious ones.

Also, replace all use of strcpy, strcat, and sprintf with strlcpy,
strlcat, and snprintf respectively.

Finally, use a new macro (PATH_SIZE) for all the static sized paths
I can find. Set to 512 for now, since that's larger than the largest
old value, and big enough for me to feel comfortable calling someone
insane if they run this with a path that long.