Disable pause when running an external program.
Not sure if this is from the terminal overhaul or from the unicode
support, but passthru_thread calls rputs() which calls cp437_out()
which calls outchar() which calls check_pause().

Also, put an EINTR wrapper around waidpid()... that wasn't the
problem I was seeing, but it's a theoretically possible one.