Allow multiple inbound directories
Some systems are configured to put inbound files with a password
into separate directories than ones without. For those systems,
if the routing is configured as Direct (the default), it's likely
that some will be password-protected (ie: with an uplink) and others
won't be (ie: some rando across the planet).

In these cases, either some processing would be needed to move all
appropriate files somewhere for The Clans to process, or The Clans
needs to be able to support at least two different inbound directories.

Because I don't know how everyone has their FTN stuff configured,
I just make it so you can configure an unlimited number of inbounds
by just adding more lines to the config file. The config *utility*
does not yet support this, and only allows setting/editing the first
one.