Fix handling of .msg files from IBBS doors.
IBBS doors configured for Binkley style outbound will put the
binkley directives in the subject, expecting them to pass through
to the FLO file.

SBBSEcho v2 and v3 since a3b2f9e2d have stripped the "delete file"
indication ('^') and generated an error, refusing to send messages
with #, ~, -, !, and @.

This changes the behaviour to translate '-' to '^', '!' to '~', and
strip '@', and keep the prefix that was in the .msg file.

This fixes the issue where outbound files from IBBS games configured
for Binkley style outbound would never get deleted.

Note that in this implementation, the prefix overrides the configured
truncation value and any "FLAGS KFL" kludge lines.