Fix race condition creating .msg file, add MAIL_NONE
We now use O_EXCL when creating the .msg file to solve that common
race condition.

Also, add a new MAIL_NONE type for when a mailer isn't used and
the *.msg files don't need to be created or deleted.

Use MAIL_NONE for the Local IBBS method. This allows us to delete
a large chunk of code that only really existed to delete .msg files.

The config utility will need to be updated to support this new type
since it will make running IBBS without FTN a bit easier.