Fix error with stream cipher handling
If the last 2k "block" in a stream cipher didn't hit exactly on
a block boundary, it was written in plaintext, not encrypted, which
would be read back as gibberish.

This caused all of the bits that add a BBS entry to seem to fail
to add with stream encrypted files (ie: ChaCha20).