Add new config option StrictMsgFile
This option requires a *.msg file to exist in the netmail directory
for the file, and that message to be addressed to the correct node.

In the near future, it will verify the source address is the correct
one from the world.ndx file too, but that's not done yet.