Move StrictMsgFile into world.ndx file
To actually get security, this needs to be set league-wide.
If a league wants to support mixed FTN/non-FTN delivery and use
the security this option provides, the non-FTN nodes will need to
be configure with a mailer type (ie: not "None") and will need to
send/receive the .MSG files along with the packet data files.

This also calls System_Error() on all the various ways to shoot
yourself in the foot with this (ie: mixing with NoMSG or MailerType None)