Allow arbitrary numeric and verbal date field separator char configurability
The wizard no longer includes the ability to change the numeric date field
separating character.

The verbal field separator is configurable only/after the verbal date output
is selected. This is a little backwards since the example output is displayed
in the numeric/verbal selection dialog, but I figured that was better than
(always) prompting for a separator character that may not actually ever be
used.

The numeric date format is always displayed in the System config menu now
(even when verbabl output format is chosen) since users can only *input* dates
in numeric format.