Start the move away from curses...
The way it was being used was pretty bad, and we don't *really* need
it for what we're doing. Instead, do ANSI and UTF-8 ourselves.

So basically, instead of using curses badly, it's moved to
reimplementing curses badly.

The only bit that's really left is install.
And testing on Windows of course.