Fix Linux harder
So, it turns out that libbsd pretty much requires glibc to work as
documented (ie: fails on Alpine with musl), so instead we'll just
use IEEE Std 1003.1-2008 ("POSIX.1").

This means we can't use cfmakeraw() either, but that's fairly easy
to roll ourselves with just POSIX flags.

We'll just use the win_wrappers functions for Linux.