More make overhaul
This adds a top-level makefile that results in a runnable game, and
terrible mess if ran inside the repo.

Also supports running from outside the repo.

This also re-adds a target to build the installer, but since the
names of the binaries have the platform and flavour in them, they're
too long for the .GUM format. The format will need to be extended
to support them before that target works.