Mangle share modes to match POSIX
We can't get an exclusive lock on files that aren't opened for write,
and we can't get a shared lock on files that aren't opened for read.

If one of the above is in effect, switch SH_DENYWR/SH_DENYRW so it's
possible.

Since this isn't multinode, all these share modes really accomplish
is allowing the program to hang, but I'm leaving the locks in for now.