Fix share modes and eliminate POSIX file accesses
The SH_DENY* macros have been deprecated in favour of _SH_DENY*
ones, so we'll use those everywhere.

Also, the last argument to _fsopen() is not a set of bitflags, it's
manifest constants, so fix _fsopen() on *nix.

Finally, eliminate open()/read()/write()/close() in favour of the
C standard stream interface. It was only used in a couple places,
and there's no need to wrap those on Windows.