Since XPDEV is built with WinSDK 10, so must pktdump
or else it fails to link with error:
unresolved external symbol __ftoul2_legacy referenced in function _msclock

pktdump was building with WinSDK 7 (and the old compatible toolchain) for
Windows XP compatibilty.

I made this change in the previous commit, but only for the debug config.
Make the same change for the release builds (which the pipelines build/use).