"Fix" language file handling
The strings file has sprintf() format specifiers for the 16-bit MS-DOS
version. Since every parameter is implicitly converted to an int, this
isn't an issue for chars, shorts, int16_t, or uint16_t (the old int
and unsigned). However, some of the format specifiers want longs,
and those values are no longer a long type (they're int32_t). This
caused various numbers to be displayed incorrectly even though the
game used them correctly.

So, I went through and added casts to the type specified in the
strings file to every snprintf() that uses numbers (or chars).
This meant looking up the macro in mstrings.h, then looking at the
referenced line in string.txt... a lot.

Anyway, looks good now.