Fix serialize of last byte of bitflags in the struct clan
The destination byte was not being zeroed correctly, so the data
that was previously in the buffer would only get additional bits
set in it, it would not get any bits cleared. This effected:
Protection remaining
First Day - This is the users first day playing.
Eliminated - Used for IBBS scores
Quest Today - Went on a quest today (only one allowed)
Attended Mass - Attended Mass today (only one allowed)
Got Blessing - Got a blessing in church today (only one allowed)

As a result, these were not being cleared during maintenance, and
various things could not be done properly in the game.