Update all IBBS sends to use the "new way"
Also, don't create all the intermediate files, just write everything
to the final packet file.

While we're here, defer writing the .msg file until IBBS_Close(), which
allows multiple packets to be combined into a single file. This feature
was always there, but it was appending *after* creating the .msg file,
which was Very Bad, and so it got disabled.