Don't let user run external program that's executing its assigned event
... using 0-byte semfile: data/event.<code>.running

The second part of fix for issue #998

Kind of abusing/reusing a node status string for the error message displayed
to a user that tries to run the program while the event is running. Might want
to address that (support an error display file or other text.dat string) later.

Background events don't work for this (since we never know if/when they're
finished executing).

If an event is configured to run exclusive, but on "any node" and you run
multiple instances of sbbs (using the same data dir), then this isn't going
to work for you since the event thread deletes any stale semfiles when it
starts up.