Eliminate all the findfirst/findnext usage
This was a terrible API in DOS, and the re-implementation of the
DOS API using the Win32 and POSIX APIs was just gross.

Replace the whole thing with a function that takes a path and a
wildcard filename and returns an allocated, NULL-terminated char **
instead.

There's still a bit of cleanup around the edges needed here, but
I'll likely do that as part of a code reorganization that moves
functions into a better set of files.

The thing as a whole has not been tested on Windows yet, but the
individual functions were.