Decrypt the buffer after writing
In the majority of cases, buffers written to disk are not used again,
however, when updating the scores as part of the BBS 1 maintenance,
the current date would get written to the ipscores.dat file from the
global.

Since EncryptWrite() would now leave it encrypted, when the date was
then copied into a users clan record, it was garbage.

Just in case something like this happens anywhere else, ensure the
buffer is unencrypted when EncryptWrite() returns.

Introduced in d1693a60