And, generate/validate signatures.
We now have all the pieces to write the bundler/unbundler utilities.
The API is a bit complex though, especially around the need to
explicitly parse TLVs.  The only way I can think of off-hand to
help with this is by passing a list of types to parseTLV which walks
the whole tree recursively parsing them... which is a fairly icky
way to do it.

I'm starting to the the feeling that I'm going to be writing a
selector language for these trees... and am worried I'll take
inspiration from CSS...