Almost, just turns out that CreatePipe() is useless.
So, ReadFile() will work the same if you don't specify an OVERLAPPED
structure regardless of a handle being overlapped or not.  However,
CreatePipe() doesn't create it overlapped, so we need to create our
own CreatePipe() thing that does open the read end for overlapped.

Also, fix some bugs in the overlapped bits and now it appears to work.

Enjoy the lowered CPU utilization.