When the user has unread personal mail, give a differnet prompt during logon
... and list/view their *unread* mail only (ordered newest first)

I'm not filtering out SPAM here since that would be weird to be prompted to
read your unread mail and then see no mail listed. sbbs_t::readmail() masks
off the NOSPAM/SPAMONLY flags from the mode parameter anyway (not that I
remember why).

If the user changes the list order (! key) when reading their unread mail,
that change is not saved to their preferences.

This implements the feature request described in issue #1027, but without any
"user setting". And it applies to all (personal) mail, not just "netmail".