A blank ReadYourMailNowQ text.dat string will not defeat the Unread mail prompt
I sbbs_t::noyes() will return true if this string is blank, so we didn't need
check here now anyway.

Setting ReadYourMailNowQ in text.dat to an empty string means this prompt
won't be shown and the read messages will *not* be shown during logon.

Setting ReadYourUnreadMailNowQ in text.dat to an empty string means this
prompt won't be shown and any unread messages *will* be shown during logon
(automatically).