Eliminate the "Start with number [n]:" prompt in initial message listing
As Deuce pointed out, this prompt can be confusing (no, it's not a menu) and
really, it's redundant because you can enter a message number at the
"Reading E-mail" prompt just as easily.

Another change is that if the list is aborted (e.g. Ctrl-C hit), leave the
current message number as the last message listed.