Wrap all ClientForm->ListView->Items accesses in try/catch
... this appears to resolve the "invalid index" exceptions that could get
thrown and pop-up error dialogs in SBBSCTRL-Win32 ... finally.

It's a mystery how the ListMutex isn't successfully protecting the ListView
(the ListView->Items->Count becomes out of sync?), but if the array index
into the ListView->Item array is invalid, it must be. Pretty weird.

Also, if WaitForSIngleObject() fails (to acquire the ListView mutex), popup
an error dialog box. I have not observed any mutex acquisition failures.