Also use a timeout if we have nothing to send.
Fixes an infinite loop where there is nothing to send because the
remote does not have the address you are calling to deliver a message
for.

For example if W6RAY just joined the BBSDev FTN, and was assigned
885:1/99, and a netmail was created welcoming the BBS to the network,
but they didn't add the new address to the AKA list in scfg, binkit
would sit using 100% CPU.

Now it sets the timeout, so sits at 0% CPU for 120s then fails.

This is still likely wrong, but at least it's wrong in a much less
harmful way.