Don't write empty PATH: control line to exported echomail message
Points don't normally have any addresses to add to the PATH: line, so
(as was pointed out in the FIDOTEST echo), an empty PATH: line would be
included in exported echomail messages from points.

Although this isn't a violation of FTS-4, it is a violation of its proposed
successor: FSC-74:

  "Blank" path lines shall not be transmitted

And this caused a duplicate PATH: line to be added by FMail when processing
such messages.

This change also eliminates possibility of adding "\r\x01PATH:" not
followed by space (ASCII 32) character, which was also a violation of FSC-74.

I resisted the urge to clean-up this crufty bit of code here and made this
commit the minimal necessary change.