Convert all C source files that use JavaScript (libmozjs) to C++
C++ is a prerequisite for all modern versions of libmozjs (SpiderMonkey).

This commit is *not* updating libmozjs, it's just doing the C->C++ conversion
using the minimum number of changes to get it to compile and run as before
using MSVC Version 19.44.35217 and GCC 11.4.0. There are still some new
warnings now however.

In general, MSVC was more permissive than GCC/G++.

This commit consists of renaming of files and resovling errors and warnings:
  - More typecasts (preferring static_cast<> where possible, mostly)
  - More proper constification (of char *'s)

Detecting C++ compilation no can be used to infer "sbbs/Terminal Server" as
JSexec and all of the other servers (except the FTP server) are now C++ as
well. This required linking the (new) SFTP library into other server projects
<shrug>.

There is a concern with services.cpp and a potential (new) memory leak with
the protected_uint32_t clients member of service_t. This is now an _Atomic
type in C++ which doesn't have a copy or move constructor so manual allocation
was required and we're probably not deleting/freeing it entirely upon shutdown
or recycle. Need to follow-up on that.

After this (more C++ heavy) build is determined to be solid, we can start
tackling the next steps in upgading libmozjs (likely version 78).