logoutuserdat() uses new function fputuserdat() to reduce open/lock/r/w ops
Reduces from 4 to 1 the open/lock/read/write/close operatons of the userbase
(user.tab) file upon user log-off.

There's more places in sbbs where we do a lot of putuser* calls that could be
(now) optimized in a similar fashion, but logoutuserdat() is called often
(especially by the web, ftp, mail and services servers), so this one was
obvious place to reduce contention on the userbase file.