Fix potential deadlock in specific case
If the telnet client sends an IP address as a Location, and that
IP address is in a filter file, the input thread mutex would not
be unlocked.

This oddball feature was suggested on Reddit here:
https://www.reddit.com/r/synchronet/comments/khuyey/ftelnetwebsocketservicejs/

And implemented in Synchronet here:
https://gitlab.synchro.net/main/sbbs/-/commit/2bbe824c0656176d0314f2d48558b322a69b9e32

It appears that it's been broken the whole time.

Found by some LLM (ChatGPT?) and reported by nelgin.