Disable TLS session cache
It appears that when the session cache is used, we don't get a copy
of the remote certificate.  Since it's now an error to not get that
copy, this would result in errors when making multiple queries to
the same TLS server in a row.

Presumably the cause of the -43 errors in letsyncrypt.