Add some passthru thread debug log messages
To hopefully help identify the cause of issue #1038

The theory being that the client socket was disconnected while running an
external program (sexyz in this case) and this check at the end of external()
(the *nix version) might have a race condition with the passthru thread
terminating due to disconnection as well:
		if (!(mode & EX_STDIN)) {
			if (passthru_thread_running)
				passthru_socket_activate(false);
			else
				pthread_mutex_unlock(&input_thread_mutex);
		}
in which case it would try to unlock the input_thread_mutex that it did not
own. I'm not clear why that would cause the pthread_mutex_destroy() call to
fail (on input_thread_mutex) but maybe it does.